requirements.txt
flask
azure-ai-formrecognizer
python-dotenv

